+++ src/main/java/com/wcs/wcslib/vaadin/widget/multifileupload/ui/UploadStatePanel.java
@@ -79,7 +79,7 @@ public class UploadStatePanel extends Panel implements MultiUploadHandler {
             String formattedErrorMsg = UploadUtil.getSizeErrorMessage(
                     multiUpload.getSizeErrorMsg(),
                     multiUpload.getMaxFileSize(),
-                    (int) event.getContentLength(),
+                    event.getContentLength(),
                     event.getFileName());
             Notification.show(formattedErrorMsg, Notification.Type.WARNING_MESSAGE);
             return false;
